requirements.txt
# PDF Agent FastAPI Requirements

# Web Framework
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Supabase
supabase>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0

# PDF Processing
PyPDF2>=3.0.0
pdf2image>=1.16.0
reportlab>=4.0.0
pillow>=10.0.0

# LLM Providers
anthropic>=0.18.0
openai>=1.0.0
google-generativeai>=0.3.0

# HTTP Client
httpx>=0.25.0
requests>=2.31.0

# Data Processing
pydantic>=2.0.0
python-dotenv>=1.0.0

# OCR (optional - for PDF OCR fallback)
# pytesseract>=0.3.10

# Utilities
aiofiles>=23.0.0
json-repair>=0.25.0
